Feature/api docs (#79)
Generate `API.md` by calling `npm run docs`.

This uses react-docgen.

Due to the default resolver of react-docgen not picking up Glamorous components, I've had to wrap some Glamorous only components with a JSX wrapper. I'd rather not have to do this but not sure of any alternative solution.

https://github.com/reactjs/react-docgen/issues/256

I'm also not sure the format of the current API.md is useful but opening PR for consideration.