Move devDependencies to root (#214)
Fixes #175 (hopefully)

Updates to a devDependency for a package causes `lerna publish` to republish that package. This has a knock on effect throughout our repo meaning a single change to one component can result in every component being republished. We don't want this to happen so are including most devDependencies at the project root (where we expect these dependencies are only used when using the package as part of a monorepo).

Exceptions are where a component has specific peerDependencies and we want to also represent that in the devDependencies.

Related: 

lerna/lerna#1269
lerna/lerna#1345
benmosher/eslint-plugin-import#935