Adjust width spacing scale
When working on creating a main layout container I noticed that the blue bar underneath the topNav was not the same as GOVUK. By looking at the source I can see the spacing scale was being used once with Calc, instead of being applied to margin-left & right, which would mean the value is unintentionally halved. This change corrects this by using a spacing scale which is double the value.