WIP: Add Storybook options
This is an attempt to enable Storybook options so we can at least have it display its stories (and sections) in alphabetical order. Then perhaps we can create a default story which will display first and could act as the readme. So far not working! See https://github.com/storybooks/storybook/issues/548#issuecomment-355278510