Refactored onClick logic
An event.preventDefault was placed in the onClick props in the Tabs.Tab hyperlink dom node. This logic has been removed out of the deliverable component. An example of how this the onClick functionality could be used by a user of the library to re-create the tabs function of govuk-fronted has been included in the fixtures using window.matchMedia. Also, some small style changes, file renames and a small addition to the test to mock the window object