Added dequote function in api-docs
To handle the different ways path.resolve handles inputs from the command line and how the command line accepts strings inputs is different windows against unix-like operating systems, a dequote function has been added. In the top-level package.json, the inputs that docs-component receives are enclosed in single quotation marks. This is now removed in the code so that path.resolve does not receive a string value with quotation marks in on windows systems.