Set storybook to be based on light theme
If your OS (such as macOS) is set to use the dark theme, some browsers pass this through to Storybook which will then use the dark theme as default. This means components will then not have a white background when viewing within Storybook.

This commit set the base theme to be the 'light' Storybook theme, regardless of what your OS setting is, as the components are meant to be placed on a white background.